v0.07 - Defining encryption and adding more functions
This time we define some more functions that we will use to work with the changes and configurations that the social network user defines. We also added two functions to work with encryption and decryption using RSA and OPENSSL in a very simple way, we created a 16-bit key to control the encryption of the information being registered in the bank. We have also created a new page where the master user or administrator can control the access and use of users on the social network.