v0.11 - Defining some parameters
This time, we define some more parameters of the functions that we will use to manipulate the database and handle user data. Soon we will implement some other important functions for the project, such as finalizing the form validation functions that will ensure that the data entered by users are in accordance with the system guidelines.