v0.11 - Working with layoult
This time we are back to working on page_home, which is where the user can see posts from friends' timelines. We add elements to the left panel tab where some components will be, such as the pages the user follows and so on. We also made a small change to the login page to make it more like the theme we chose for the project.